requirements.txt
colorama
pytz_deprecation_shim
croniter
discord.py
ciso8601
psutil
pyaml